meson: add wrap for libdrm
This allows building Mesa with the Android NDK, which doesn't provide
libdrm.

We will generate an `Android.bp` file using the `ninja-to-soong` tool
(https://github.com/rjodinchr/ninja-to-soong), and to do this we run
Mesa's standard meson build system to generate the ninja commands that
we then translate to soong.

That meson invocation is done using the Android NDK, which doesn't
provide libdrm, so until we find an alternative solution[*] we provide a
wrap file that builds libdrm as part of the Mesa build (but does not
install it--we still use the Android-provided libdrm at runtime)

* We will likely add support to `ninja-to-soong` to generate
  `Android.bp` for libdrm. Once done, we'll drop this patch from our
  tree.

Bug: b/393645057
Test: Build Mesa outside the AL tree using the Android NDK
Flag: EXEMPT desktop only
Change-Id: Ibee53454c473afe2fc561b2624d9d1705091c53a